revision comments, object views separated
