+++ backend/app/main/answer_views.py
@@ -1,0 +1,40 @@
+from flask import url_for, session, request
+from twilio.twiml.messaging_response import MessagingResponse
+
+from . import main
+from .. import db
+from .models import Answer, Question
+
+
+@main.route('/answer/<question_id>', methods=['POST'])
+def answer(question_id):
+    question = Question.query.get(question_id)
+
+    db.save(Answer(content=request.values['Body'],
+                   question=question,
+                   session_id=session['id']))
+
+    next_question = question.next()
+    if next_question:
+        return redirect_twiml(next_question)
+    else:
+        return goodbye_twiml()
+
+
+# Redirect to the question route
+def redirect_twiml(question):
+    response = MessagingResponse()
+    response.redirect(url=url_for('main.question', question_id=question.id),
+                      method='GET')
+    return str(response)
+
+
+# Compose end of survey text
+# TODO show all entered values and ask user to confirm
+def goodbye_twiml():
+    response = MessagingResponse()
+    response.message("Thank you for completing the signup form! We look forward to working with you!")
+    if 'question_id' in session:
+        del session['question_id']
+    return str(response)
+
