+++ backend/config.py
@@ -1,3 +1,5 @@
+import os
+
 class Config:
     APP_NAME = "Move Up CRM Tool"
     SECRET_KEY = os.environ.get("SECRET_KEY") or "hard to guess string"
